Added Outlet, overrides for Fans and Outlets
This patch removes the “dimmertest” as the dimmers seem to work fine.
It provides an “Outlet” type. This can be configured in the Home app to be a Light, a Fan, or an Outlet.

In order to declare a device an Outlet, use config.js
“Outlets”: [123,124,125],
this will override the device type to outlet or these IDs (remember: these are Vera IDs as displayed by the plugin, not ZWave IDs).

For completeness, the patch also includes a “Fans” override (this is for fans that can get/set rotation speed. I don’t have one so it’s untested.